+++ controllers/apidefinition_controller.go
@@ -176,6 +176,7 @@ func (r *ApiDefinitionReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 			}
 		}
 
+		// Check GraphQL Federation
 		if upstreamRequestStruct.Spec.GraphQL != nil && upstreamRequestStruct.Spec.GraphQL.GraphRef != "" {
 			if upstreamRequestStruct.Spec.GraphQL.ExecutionMode == model.SubGraphExecutionMode {
 				subgraph := &tykv1alpha1.SubGraph{}
@@ -188,8 +189,8 @@ func (r *ApiDefinitionReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 					return err
 				}
 
-				upstreamRequestStruct.Spec.GraphQL.Schema = subgraph.Spec.Subgraph.Schema
-				upstreamRequestStruct.Spec.GraphQL.Subgraph.SDL = subgraph.Spec.Subgraph.SDL
+				upstreamRequestStruct.Spec.GraphQL.Schema = subgraph.Spec.Schema
+				upstreamRequestStruct.Spec.GraphQL.Subgraph.SDL = subgraph.Spec.SDL
 
 				subgraph.Status.APIID = upstreamRequestStruct.Spec.APIID
 				err = r.Status().Update(ctx, subgraph)
@@ -212,19 +213,19 @@ func (r *ApiDefinitionReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 				for _, ref := range supergraph.Spec.SubgraphRefs {
 					ns := namespaceValue(ref.Namespace, supergraph.Namespace)
 
-					sg := &tykv1alpha1.SubGraph{}
+					subGraph := &tykv1alpha1.SubGraph{}
 					err := r.Client.Get(ctx, types.NamespacedName{
 						Name:      ref.Name,
 						Namespace: ns,
-					}, sg)
+					}, subGraph)
 					if err != nil {
 						return err
 					}
 
-					ns, name := decodeID(sg.Status.APIID)
+					ns, name := decodeID(subGraph.Status.APIID)
 
-					apidef := &tykv1alpha1.ApiDefinition{}
-					err = r.Client.Get(ctx, types.NamespacedName{Namespace: ns, Name: name}, apidef)
+					apiDef := &tykv1alpha1.ApiDefinition{}
+					err = r.Client.Get(ctx, types.NamespacedName{Namespace: ns, Name: name}, apiDef)
 					if err != nil {
 						return err
 					}
@@ -232,14 +233,15 @@ func (r *ApiDefinitionReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 					upstreamRequestStruct.Spec.GraphQL.Supergraph.Subgraphs = append(
 						upstreamRequestStruct.Spec.GraphQL.Supergraph.Subgraphs,
 						model.GraphQLSubgraphEntity{
-							APIID: sg.Status.APIID,
-							Name:  apidef.Spec.Name,
-							URL:   fmt.Sprintf("tyk://%s", apidef.Name),
-							SDL:   sg.Spec.Subgraph.SDL,
+							APIID: subGraph.Status.APIID,
+							Name:  apiDef.Spec.Name,
+							URL:   fmt.Sprintf("tyk://%s", apiDef.Name),
+							SDL:   subGraph.Spec.SDL,
 						})
 				}
 
 				upstreamRequestStruct.Spec.GraphQL.Schema = supergraph.Spec.Schema
+				upstreamRequestStruct.Spec.GraphQL.Supergraph.MergedSDL = supergraph.Spec.MergedSDL
 			}
 		}
 
@@ -261,36 +263,6 @@ func (r *ApiDefinitionReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		}
 
 		upstreamRequestStruct.Spec.CollectLoopingTarget()
-
-		if upstreamRequestStruct.Spec.GraphQL != nil {
-			if upstreamRequestStruct.Spec.GraphQL.Enabled {
-				switch upstreamRequestStruct.Spec.GraphQL.ExecutionMode {
-				case "subgraph":
-					// Update subgraph's status field with the introspection results. However, introspection has not been
-					// implemented yet. Therefore, we are populating the subgraph's status with the SDL given manually.
-
-					// TODO: implement introspection and update following code piece. So, the operator gets the subgraph's
-					// SDL from the introspection result, which will be used in the supergraph's
-					// spec.graphql.supergraph.subgraphs field.
-					if upstreamRequestStruct.Status.SDL !=
-						upstreamRequestStruct.Spec.GraphQL.Subgraph.SDL {
-						upstreamRequestStruct.Status.SDL = upstreamRequestStruct.Spec.GraphQL.Subgraph.SDL
-					}
-				case "supergraph":
-					// Check if the supergraph's merged_sdl is valid or not. If the dashboard is in use, merged_sdl might
-					// be empty because it is the as spec.graphql.schema. However, if the Tyk CE is in use, merged_sdl
-					// should be provided.
-					if upstreamRequestStruct.Spec.GraphQL.Supergraph.MergedSDL == "" {
-						if env.Mode == "ce" {
-							// if Tyk CE gateway is used, merged_sdl should be provided through ApiDefinition YAML file.
-							return fmt.Errorf("expected to have merged_sdl declared for Tyk CE")
-						}
-
-						upstreamRequestStruct.Spec.GraphQL.Supergraph.MergedSDL = upstreamRequestStruct.Spec.GraphQL.Schema
-					}
-				}
-			}
-		}
 
 		//  If this is not set, means it is a new object, set it first
 		if desired.Status.ApiID == "" {
@@ -700,11 +672,11 @@ func (r *ApiDefinitionReconciler) updateStatus(
 	return r.Status().Update(ctx, &api)
 }
 
-func (r *ApiDefinitionReconciler) findGraphsForApiDefinition(apiDef client.Object) []reconcile.Request {
+func (r *ApiDefinitionReconciler) findGraphsForApiDefinition(graph client.Object) []reconcile.Request {
 	apiDefDeployments := &tykv1alpha1.ApiDefinitionList{}
 	listOps := &client.ListOptions{
-		FieldSelector: fields.OneTermEqualSelector(GraphKey, apiDef.GetName()),
-		Namespace:     apiDef.GetNamespace(),
+		FieldSelector: fields.OneTermEqualSelector(GraphKey, graph.GetName()),
+		Namespace:     graph.GetNamespace(),
 	}
 
 	if err := r.List(context.TODO(), apiDefDeployments, listOps); err != nil {